[smart-release #164] improve handling of empty commits
This allows to detect if empty commits are possible by checking that
manifests actually changed. If not, we create empty commits to
have something specific to tag.

Note that we don't actually detect changes, so Cargo.lock changes
might still be possible. For now, allow empty commits.

One day this can be better by reusing a previous commit.