repository(fix): loose reference iteration with non-dir prefixes… (#198)
Previously it was expected for the prefix `Path` to always exist for
the prefix to be valid. This, however, is not similar to packed
prefixes, which allow non-dir prefixes as well.

Now we will check if the prefix is actually a directory, and if not
split it into its parent directory and the filename portion. The latter
is then used for prefix matching file names within that directory.