fix: build complete history information to match with `did crate changed` queries
Previously it was possible see a crate was changed, but didn't receive a
version bump which would in turn halt the release process.

The issue was an algorithm inabuility to find changes in the commitgraph
because it would not look at the correct tree, causing trees to be
missed entirely. This in turn caused it to not see changes that were
present and the mismatch in question.