feat: Recognize variants of "thanks clippy"
The commit message title (or title prefix) "thanks clippy" was
recognized, but case variants such as "Thanks clippy" were not.
Less importantly, the plausible "thanks, clippy" variant (and its
case variants) were also not recognized.

This downcases the title before checking for "thanks clippy" as a
prefix, and it adds a check for the prefix "thanks, clippy".

Fixes #39