Make the main CI workflow easier to run in forks
This adds patterns besides the literal name `main` so CI runs on
`push` to any branch that has `run-ci` as a non-trailing component.
It also adds a `workflow_dispatch` trigger so the workflow can be
run (on any branch) manually from the Actions tab.

These are to make the workflow easier to run in forks, roughly
analogous to https://github.com/GitoxideLabs/gitoxide/pull/1507.