Enable Dependabot version updates for actions
This enables grouped Dependabot version updates, for GitHub Actions
only (not for Rust dependncies).

Currently some actions in use in this project's workflows specify
old major versions that are not currently maintained with bug fixes
and that use old unsupported Node.js version.

This includes an old version of `actions/checkout` (using `v3`,
where `v4` is needed to ensure a supported version of Node.js is
used) and SARIF-related actions. On the latter, see:
https://github.blog/changelog/2025-01-10-code-scanning-codeql-action-v2-is-now-deprecated/

Under this new configuration, Dependabot will open a single pull
request to upgrade all actions that have available updates,
followed by pull request monthly or less for this in the future.