Bump `gix` from 0.69.1 to 0.71.0
This upgrades `gix-*` crates other than as depended on transitively
through `gix-testtools`.

Crate versions are doubled up, due to `gix-testtools` depending on
a major version lower, but this was already largely the case.

The upgrade decreases the use of dependency versions affected by
RUSTSEC-2025-0021, now only using them in the test suite.

Although the upgrade does not break anything or make tests fail,
this also adapts to the change by replacing a call to the now
deprecated `Repository::work_dir()` method with the non-deprecated
equivalent `Repository::workdir()`.