Find all applicable RUSTSEC advisories in `security_audit` job
This works around a bug in the `actions-rs/cargo-check` action,
https://github.com/actions-rs/audit-check/issues/163, where it runs
`cargo generate-lockfile` even if a fully usable `Cargo.lock` is
already present, before running `cargo audit`, thus missing any
advisories that affect locked dependency versions but not the
latest usable versions that updating `Cargo.lock` would change to.

The action runs `cargo generate-lockfile` and cannot be configured
not to do so without modifying the action code itself. It also
checks that command's exit status, failing fast if it failed
(otherwise `chmod -w Cargo.lock` or `sudo chattr +i Cargo.lock`
could be workable solutions). Fortunately, because we are only
running this on GitHub-hosted runners, which regenerate the whole
virtual machine each time, it is okay to modify the behavior of
`cargo` itself.

This renames the `cargo` binary to `cargo.orig` and replaces it
with a wrapper script named `cargo` that treats `generate-lockfile`
as a special case and does nothing other than display a warning
that it is not running the command. Otherwise it hands off to
`cargo.orig`, such as to let `cargo audit` itself run in full.

Neither the general approach of wrapping `cargo` to stub out
`generate-lockfile`, nor the specific details of how that is
achieved here, are implemented here in a fully general way. The
approach here is not sufficiently robust for general use, but it
should be more than good enough until the use of the unmaintained
`actions-rs/cargo-check` action is eventually replaced with some
other way to report advisories in an easy-to-read manner.

The squashed history of a few commits, which led to this approach,
are shown below. For their full details, see EliahKagan#4.

* Investigate cargo-audit behavior on CI

* See what happens if `Cargo.lock` is marked read-only

* Wrap `cargo` to stub out `generate-lockfile`

* Clarify and clean up comments

- Expand and clarify the comment explaining why assuming `$1` is
  the subcommand rather than an option is expected to work in the
  `cargo` wrapper script.

- Clean up the already commented-out step that had been used to run
  a reverse shell for debugging.