Bump locked dependencies
This runs `cargo update` to fix #50.

Although various non-`gix-*` crates are included among those
updated, the main anticipated impact of this change is to make it
so the versions of `gix-features` and associated `gix-*` crates
used as non-dev dependencies have the fix for RUSTSEC-2025-0021.

For more details, see #50 and:
https://gist.github.com/EliahKagan/b9591e61ae94c3680ffa20a20d384efe