Thanks clippy
This removes a call to `to_string()` that is no longer necessary:

    warning: `to_string` applied to a type that implements `Display` in `trace!` args
       --> src/command/release/manifest.rs:568:52
        |
    568 | ...                   current_version_req.to_string()
        |                                          ^^^^^^^^^^^^ help: remove this
        |
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#to_string_in_format_args
        = note: `#[warn(clippy::to_string_in_format_args)]` on by default

Because `cargo clippy` is run on CI with `--allow deprecated`, this
is enough to make it pass on CI. However, running `cargo clippy`
without `--allow deprecated` also complains about the use of the
deprecated type alias `winnow::PResult`. That other warning is also
new since the immediately preceding run of `cargo update`. This
commit does not fix that warning.