Thanks clippy
This adapts to `winnow` changes. Specifically, `changelog.parse`
was using `PResult`, which is deprecated since `winnow` 0.6.25:

https://github.com/winnow-rs/winnow/blob/main/CHANGELOG.md#0625---2025-01-27

This changes from `PResult` to `ModalResult`, which is its
recommended replacement.

At this time, with the current version, `PResult` is an alias for
`ModalResult`. So while this does not establish that the version
upgrade itself hasn't caused unwanted behavioral changes, replacing
`PResult` with `ModalResult` given that we are at this version does
not cause any such problem.

The `clippy` warning this fixes is:

    warning: use of deprecated type alias `winnow::PResult`: Replaced with ModalResult
       --> src/changelog/parse.rs:473:95
        |
    473 | fn headline<'a, E: ParserError<&'a str> + FromExternalError<&'a str, ()>>(i: &mut &'a str) -> PResult<Headline...
        |                                                                                               ^^^^^^^
        |
        = note: `#[warn(deprecated)]` on by default